Store diff info as buffer-local variables
Refactored diff handling to store all diff-related information as
buffer-local variables in the diff buffer itself, instead of relying
on hash table lookups. This simplifies the code and ensures data
locality - the information stays with the buffer it belongs to and
is automatically cleaned up when the buffer is killed.

Changes:
- Add buffer-local variables to diff buffer for all diff info
- Simplify monet--cleanup-diff to use buffer-local variables
- Update monet--close-tab to get info from buffer-local variables
- Update monet--handle-diff-accept to use buffer-local variables